Add detail

+++ model/Hackathon.cpp
@@ -138,10 +138,10 @@ model::Hackathon::Hackathon(const Json::Value& value) {
     this->sponsor = value["sponsor"].asString();
     this->reward = value["reward"].asString();
     struct tm begin_date;
-    begin_date.tm_sec = std::stoi(value["begin_date"].asString());
+    begin_date.tm_sec = std::stol(value["begin_date"].asString());
     this->begin_date = mktime(&begin_date);
     struct tm ended_date;
-    ended_date.tm_sec = std::stoi(value["ended_date"].asString());
+    ended_date.tm_sec = std::stol(value["ended_date"].asString());
     this->ended_date = mktime(&ended_date);
     for(auto iterator = value["steps"].begin();iterator != value["steps"].end(); ++iterator) {
         this->steps.push_back(Step((*iterator)));
